requirements/runtime.txt
boto3==1.34.160
cryptography==42.0.8
jsonschema==4.23.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
pika==1.3.2
prometheus_client==0.20.0
psycopg[binary]==3.2.1
PyJWT==2.8.0
PyYAML==6.0.2
